requirements.txt
attrs==19.3.0
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
cffi==1.13.2
constantly==15.1.0
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.9
django-bootstrap-ui==0.5.1
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-filter==2.2.0
django-tables2==2.2.1
django-tag-parser==2.1
docopt==0.6.1
dominate==2.1.17
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
paramiko==2.7.1
pexpect==3.0
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
Pygments==2.5.2
PyHamcrest==2.0.0
PyNaCl==1.3.0
pytz==2019.3
setproctitle==1.1.8
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
termcolor==1.1.0
Twisted==19.10.0
zope.interface==4.7.1
whitenoise
gunicorn
